Add further functionality around the game mode concept.
Switch Component is now in a new TopHeader component. When the switch is
toggled, so is the mode (from regular to Devlandia, and back).

When the mode is toggled, most of the app reacts to it,  whether it means a
change to the background color, text color or SVG colors.